requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
openai==1.0.0
pymupdf==1.22.4        
opencv-python-headless 
ezdxf                 
python-docx
reportlab
flask-cors
psycopg2-binary
PyMuPDF